Initial implementation of custom "alt style"
This is a work-in-progress that will live on a separate branch.  I am
not sure whether it will be merged into 'master', because there may be
some cases where these slightly accented backgrounds do not combine well
with other accented backgrounds that assumed a greyscale backdrop.  My
tests suggest that such cases should not pose any major hindrance.

In terms of colour contrast, all these alternative base values conform
with the highest accessibility standard of minimum 7:1 contrast relative
to all foreground values that are meant to be combined with them.

All new colour values are tentative.  Expect further tweaks to make
things more consistent.  Furthermore, there are more variables that need
to be covered, so that all base greyscale value adapt accordingly.